update test data with more recent dates
Unfortunately, with the recent changes to how lost partners are calculated,
the output is sensitive to the current time.
Without this change, the test data was resulting in all of the partners being lost,
which makes it not very useful.

Until we figure out how to nicely mock the current time when appropriate,
we'll just have to update the test data with new 'lastTransactionDate' dates
every year or so.